Update FB_TempSensor.TcPOU
Guard against div by zero. I ran out of ideas for what to set temp or scale to in the case of a zero scale. You can alternatively use a MAX function and pick a minimum value for iScale to default to if it's below that level. eg. MAX(iScale, 1E-9)