I had to redo the auto-reset thing... it still didn't make sense.
Turns out the test I had for it was passing due to some other erroneous condition. Once I isolated this test it became obvious the autoreset was still non-functional.

Reviewing the logic here I also decided that the decaying/countdown element didn't make much sense either... It basically amounts to extra conditions for a reset, that are redundant and confusing.

To keep things simpler and understandable I want to have one criteria for auto-reset: the log message is behaving well now, and has proved it behaves well for a little while (like 10 seconds or so).